requirements.txt
certifi==2019.11.28
chardet==3.0.4
cvxpy==1.0.25
cycler==0.10.0
Cython==0.29.14
dill==0.3.1.1
ecos==2.0.7.post1
future==0.18.2
idna==2.8
joblib==0.14.1
kiwisolver==1.1.0
matplotlib
multiprocess==0.70.9
numpy==1.18.1
opt-einsum==3.1.0
osqp==0.6.1
Pillow==7.0.0
pyparsing==2.4.6
python-dateutil==2.8.1
python-mnist==0.6
qutip==4.4.1
requests==2.22.0
scipy==1.4.1
scs==2.1.1.post2
six==1.13.0
torch==1.3.1
torchvision==0.4.2
tqdm==4.41.1
urllib3==1.25.7
